+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Date format: DD/MM/YYYY
 - `PaneItemAction.body` is no longer required ([#545](https://github.com/bdlukaa/fluent_ui/issues/545))
 - Added `DropDownButton.onOpen` and `DropDownButton.onClose` callbacks ([#437](https://github.com/bdlukaa/fluent_ui/issues/537))
 - Ensure `MenuFlyoutItem.onPressed` is called after the flyout is closed if `DropDownButton.closeAfterClick` is true ([#520](https://github.com/bdlukaa/fluent_ui/issues/520))
+- Ensure the `TimePicker` and `DatePicker` popups will fit if the screen is small ([#544](https://github.com/bdlukaa/fluent_ui/issues/544))
 
 ## 4.0.0